Cleaned up some global_vars imports
