+++ flika/__init__.py
@@ -2,10 +2,9 @@
 Flika: An interactive image processing program for biologists written in Python.
 """
 
-from flika.version import __version__
-from flika.logger import logger
-from flika.flika import start_flika
 import flika.global_vars as g
+from flika.flika import start_flika
+from flika.version import __version__
 
 # Define public API
 __all__ = ["start_flika", "__version__", "g"]
